Make interface for PedestrianTracker
